Reduced duplication of code.
