+++ src/exceptions/InternalBinaryOperationException.php
